[Tests] Refactor unit tests structure
