+++ tests/projects/assets/test_kfp_input_file.txt
